refactor(web): move the banner decision out of the view, and read null the same way twice
Three review points.

`state: null` was refused while `lowBalanceMicro: null` was tolerated — the same
inconsistency, in a file whose whole job is boundary discipline, and `null` is the natural
serialization if the service ever models "no gateway configured" explicitly. Both now read
`!= null`. `balanceBanner` already did the right thing with it for free: null matches
neither arm, so it falls through and the page claims nothing, exactly as for an older
service — while the balance rules still apply, since those owe nothing to the gateway.

The low-balance guard went back to `> 0`. Truthiness catches 0, null and undefined but a
negative threshold is truthy, and `assertAccount` accepts any finite number. Reaching the
comparison with one takes an overspent balance under a negative threshold, which is absurd
but not unrepresentable — and the first test I wrote for it did not actually exercise the
bug, since `1 < -5_000_000` is false whatever the guard says. It now uses a balance below
the negative threshold, and removing the guard reddens it.

`balanceBanner` moved to `lib/billing-banner.ts`. It is pure over a plain object, but its
test imported the view and with it `next/navigation`, `swr`, `org-context` and the UI kit —
the only test in the package importing that module without happy-dom and without mocking
navigation. The test now imports one function, the env question is gone, and the decision
is no longer part of a 'use client' module's exported surface.
